+++ src/App.vue
@@ -1,43 +1,75 @@
 <script setup lang="ts">
-import HelloWorld from "./components/HelloWorld.vue";
-import DragBall from "./components/DragBall.vue";
-import SearchDrawer from "./components/SearchDrawer.vue";
+import HelloWorldVue from "./components/HelloWorld.vue";
+import DragBallVue from "./components/DragBall.vue";
+import MainDrawerVue from "./components/MainDrawer.vue";
 
-import { ref, provide } from "vue";
+import { ref, provide, reactive, watch } from "vue";
+import { IConfig } from "./types/config";
+import { SiyuanClient } from "./utils/siyuan";
 
 const visible = ref(false); // 抽屉是否可见
+const size = ref(0.5); // 抽屉宽度占比
+
+const config: IConfig = reactive({
+    server: {
+        protocol: "http",
+        hostname: "localhost",
+        port: 6806,
+        token: "",
+        url: new URL("http://localhost:6806"),
+    },
+});
+const client = new SiyuanClient(config.server.url, config.server.token);
+
+watch(
+    () => config.server,
+    newVal => {
+        config.server.url.protocol = newVal.protocol;
+        config.server.url.hostname = newVal.hostname;
+        config.server.url.port = String(newVal.port);
+
+        client.update(config.server.url, config.server.token);
+    },
+    {
+        immediate: true, // 立即执行一次
+        deep: true, // 深层跟踪
+    },
+);
+
 // REF: [依赖注入 | Vue.js](https://cn.vuejs.org/guide/components/provide-inject.html#provide)
 provide("visible", visible);
+provide("config", config);
+provide("client", client);
 </script>
 
 <template>
-    <HelloWorld msg="Vite + Vue" />
-    <!-- 打开抽屉的悬浮球 -->
-    <DragBall
-        style="z-index: 1"
-        :top="'2em'"
-        :right="'2em'"
-    >
-        <a-button
-            type="outline"
-            shape="circle"
-            @click="visible = !visible"
-        >
-            <img
-                style="width: 75%; height: 75%"
-                src="./assets/siyuan-32.png"
-                alt="siyuan logo"
-            />
-        </a-button>
-    </DragBall>
-    <!-- 抽屉 -->
+    <HelloWorldVue msg="Vite + Vue" />
     <Teleport to="body">
+        <!-- 打开抽屉的悬浮球 -->
+        <DragBallVue
+            style="z-index: 1"
+            :top="'2em'"
+            :right="'2em'"
+        >
+            <a-button
+                type="outline"
+                shape="circle"
+                @click="visible = !visible"
+            >
+                <img
+                    style="width: 75%; height: 75%"
+                    src="./assets/siyuan-32.png"
+                    alt="siyuan logo"
+                />
+            </a-button>
+        </DragBallVue>
         <!--
             REF [Arco Design Vue](https://arco.design/vue/component/split)
             分割面板, 用于调整抽屉的宽度
-         -->
+        -->
         <a-split
             class="split-container"
+            v-model:size="size"
             v-show="visible"
         >
             <template #first>
@@ -56,9 +88,10 @@ provide("visible", visible);
                     class="split-panel"
                 >
                     <!-- 抽屉内容 -->
-                    <SearchDrawer
-                        :width="'100%'"
+                    <MainDrawerVue
                         style="z-index: unset; position: unset"
+                        :width="'100%'"
+                        :footer="false"
                         popupContainer="#drawerContainer"
                     />
                 </a-layout>
@@ -69,7 +102,7 @@ provide("visible", visible);
 
 <style scoped lang="less">
 .split-container {
-    z-index: 2147483647;
+    z-index: 1000;
     position: fixed;
     top: 0;
     left: 0;