+++ src/components/TabSearch.vue
@@ -1,0 +1,10 @@
+<!-- REF [Arco Design Vue](https://arco.design/vue/component/tabs) -->
+<script setup lang="ts"></script>
+
+<template>
+    <a-tab-pane key="1">
+        <template #title> <icon-search /> {{ $t("search_result") }} </template>
+    </a-tab-pane>
+</template>
+
+<style scoped lang="less"></style>