+++ src/locales/en.ts
@@ -1,5 +1,24 @@
 export default {
-    siyuan: "SiYuan",
-    settings: "Settings",
+    client: "Client",
+    conect_siyuan_client: "Connect @:siyuan @:client",
+    conect_success: "Connect @:success",
+    help: {
+        server: "@:siyuan_server URL Origin",
+        token: "Available in @:path.token"
+    },
+    hostname: "IP Address {'|'} Domain Name {'|'} Host Name",
+    path: {
+        about: "About",
+        settings: "Settings",
+        token: "@:path.settings > @:path.about > @:token"
+    },
     search: "Search",
+    search_result: "Search Results",
+    server: "Server",
+    settings: "User Settings",
+    siyuan: "SiYuan",
+    siyuan_server: "@:siyuan@:server",
+    success: "Success",
+    test: "Test",
+    token: "API Token"
 };