+++ src/locales/zh-Hans.ts
@@ -1,5 +1,24 @@
 export default {
-    siyuan: "思源",
-    settings: "设置",
+    client: "客户端",
+    conect_siyuan_client: "连接@:siyuan@:client",
+    conect_success: "连接@:success",
+    help: {
+        server: "@:siyuan_server URL 源",
+        token: "可在 @:path.token 中获取"
+    },
+    hostname: "IP 地址 {'|'} 域名 {'|'} 主机名",
+    path: {
+        about: "关于",
+        settings: "设置",
+        token: "@:path.settings > @:path.about > @:token"
+    },
     search: "搜索",
+    search_result: "搜索结果",
+    server: "服务",
+    settings: "用户设置",
+    siyuan: "思源",
+    siyuan_server: "@:siyuan@:server",
+    success: "成功",
+    test: "测试",
+    token: "API Token"
 };