+++ src/locales/zh-Hant.ts
@@ -1,5 +1,24 @@
 export default {
-    siyuan: "思源",
-    settings: "設置",
+    client: "客戶端",
+    conect_siyuan_client: "連接@:siyuan@:client",
+    conect_success: "連接@:success",
+    help: {
+        server: "@:siyuan_server URL 源",
+        token: "可在 @:path.token 中獲取"
+    },
+    hostname: "IP 地址 {'|'} 域名 {'|'} 主機名",
+    path: {
+        about: "關於",
+        settings: "設置",
+        token: "@:path.settings > @:path.about > @:token"
+    },
     search: "搜索",
+    search_result: "搜索結果",
+    server: "服務",
+    settings: "用戶設置",
+    siyuan: "思源",
+    siyuan_server: "@:siyuan@:server",
+    success: "成功",
+    test: "測試",
+    token: "API Token"
 };