+++ src/types/config.d.ts
@@ -1,0 +1,9 @@
+export interface IConfig {
+    server: {
+        protocol: string;
+        hostname: string;
+        port: number;
+        token: string;
+        url: URL;
+    };
+}